backend/requirements.txt
aiodns==1.1.1
aiohttp==2.3.2
aiopg==0.13.1
aioredis==0.3.5
amqp==2.2.2
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
billiard==3.5.0.3
cchardet==2.1.1
celery==4.1.0
certifi==2017.7.27.1
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
defusedxml==0.5.0
Django==1.11.5
django-braces==1.11.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-redis==4.8.0
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.4
dnspython==1.15.0
dnspython3==1.15.0
hiredis==0.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
Jinja2==2.9.6
kombu==4.1.0
liqpay-python3==1.1.1
Markdown==2.6.9
MarkupSafe==1.0
model-mommy==1.4.0
msgpack-python==0.4.8
multidict==3.3.2
oauthlib==2.0.6
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pycares==2.3.0
pyIsEmail==1.3.1
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
SQLAlchemy==1.1.15
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
yarl==0.13.0
zope.interface==4.4.3
